move src/components/templates/TalkShowStudio/useStage.ts -> src/hooks/useStage.ts
